% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{emodnet_get_layers}
\alias{emodnet_get_layers}
\title{Get EMODnet WFS layers}
\usage{
emodnet_get_layers(
  wfs = NULL,
  service = "seabed_habitats_individual_habitat_map_and_model_datasets",
  service_version = "2.0.0",
  layers,
  crs = 4326,
  cql_filter = NULL,
  reduce_layers = FALSE,
  suppress_warnings = FALSE
)
}
\arguments{
\item{wfs}{A `WFSClient` R6 object with methods for interfacing an OGC Web Feature Service.}

\item{service}{the EMODnet OGC WFS service name. Defaults to the EMODnet Seabed
Habitats WFS `"seabed_habitats_individual_habitat_map_and_model_datasets"`. For
available services, see `emodnet_wfs`.}

\item{service_version}{the WFS service version. Defaults to "2.0.0".}

\item{layers}{a character vector of layer names. To get info on layers, including
`layer_name` use `emodnet_get_wfs_info()`}

\item{crs}{integer. An EPSG code for the output crs. Defaults to 4326, corresponding to `"+proj=longlat +datum=WGS84 +no_defs"`,}

\item{cql_filter}{character. Features returned can be filtered using valid Extended Common Query Language (ECQL)
filtering statements (\url{https://docs.geoserver.org/stable/en/user/filter/ecql_reference.html}). Should be one of:
 \itemize{
  \item{character string or character vector of length 1. Filter will be recycled across all layers requested}
  \item{character vector of length equal to the length of layers. Filter will be matched to layers sequentially.
  Elements containing `NA` are ignored}
  \item{named character vector. Each filter will be applied to the layer corresponding to the filter name.
  Filters with names that do not correspond to any layers are ignored. Layers without corresponding filters are returned whole }
}}

\item{reduce_layers}{whether to reduce output layers to a single `sf` object.}

\item{suppress_warnings}{logical. Whether to suppress messages of layer
download failures.}
}
\value{
If `reduce_layers = FALSE` (default), a list of `sf`
objects, one element for each layer. Any layers for which download was
unsuccessful will be NULL. If `reduce_layers = TRUE`, all layers are
reduced (if possible) to a single `sf` containing data for all layers.
`NULL` layers are ignored.
}
\description{
Performs an WFS getFeature request for layers from a `wfs` object or specified EMODnet Service. Filtering
of layer features can also be handled via ECQL language filters.
}
\examples{
emodnet_get_layers(layers = c("dk003069", "dk003070"))
emodnet_get_layers(layers = c("dk003069", "dk003070"), reduce_layers = TRUE)
emodnet_get_layers(service = "human_activities", layers = "maritimebnds",
    cql_filter = "sitename='Territory sea (12 nm)'" )
}
