% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{emodnet_wfs}
\alias{emodnet_wfs}
\title{EMODnet Web Feature Services}
\format{
A tibble of available EMODnet Web Feature Services
\describe{
\item{service_name}{Service name}
\item{service_url}{Service URL}
}
}
\usage{
emodnet_wfs
}
\description{
EMODnet Web Feature Services
}
\keyword{datasets}
