% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMODnetWFS-package.R
\docType{package}
\name{EMODnetWFS-package}
\alias{EMODnetWFS}
\alias{EMODnetWFS-package}
\title{EMODnetWFS: Access EMODnet Web Feature Service data through R}
\description{
Access and interrogate EMODnet Web Feature Service data through R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://emodnet.github.io/EMODnetWFS/}
  \item \url{https://github.com/EMODnet/EMODnetWFS}
  \item Report bugs at \url{https://github.com/EMODnet/EMODnetWFS/issues}
}

}
\author{
\strong{Maintainer}: Salvador Fernández-Bejarano \email{salvador.fernandez@vliz.be} (\href{https://orcid.org/0000-0003-0535-7677}{ORCID})

Authors:
\itemize{
  \item Anna Krystalli \email{annakrystalli@googlemail.com} (\href{https://orcid.org/0000-0002-2378-4915}{ORCID})
  \item Maëlle Salmon \email{msmaellesalmon@gmail.com} (\href{https://orcid.org/0000-0002-2815-0399}{ORCID})
}

Other contributors:
\itemize{
  \item Thomas J Webb \email{t.j.webb@sheffield.ac.uk} [contributor]
  \item European Marine Observation Data Network (EMODnet) Biology project European Commission's Directorate - General for Maritime Affairs and Fisheries (DG MARE) \email{bio@emodnet.eu} [copyright holder]
  \item VLIZ (VLAAMS INSTITUUT VOOR DE ZEE) \email{info@vliz.be} [funder]
}

}
\keyword{internal}
