% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_attributes.R
\name{layer_attribute_descriptions}
\alias{layer_attribute_descriptions}
\title{Get layer attribute description}
\usage{
layer_attribute_descriptions(
  wfs = NULL,
  service = NULL,
  service_version = "2.0.0",
  layer
)
}
\arguments{
\item{wfs}{A \code{WFSClient} R6 object with methods for interfacing an OGC Web Feature Service.}

\item{service}{the EMODnet OGC WFS service name.
For available services, see \code{\link[=emodnet_wfs]{emodnet_wfs()}}.}

\item{service_version}{the WFS service version. Defaults to "2.0.0".}

\item{layer}{character sting of layer name. To get info on layers, including
\code{layer_name} use \code{\link[=emodnet_get_wfs_info]{emodnet_get_wfs_info()}}.}
}
\value{
data.frame containing layer attribute descriptions (metadata).
}
\description{
Get layer attribute description
}
\examples{
\dontrun(
layer_attribute_descriptions(service = "human_activities", layer = "maritimebnds")
)
}
