% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_properties.R
\name{layer_attribute_inspect}
\alias{layer_attribute_inspect}
\title{Inspect layer attributes}
\usage{
layer_attribute_inspect(
  wfs = NULL,
  service,
  service_version = "2.0.0",
  layer,
  attribute
)
}
\arguments{
\item{wfs}{A `WFSClient` R6 object with methods for interfacing an OGC Web Feature Service.}

\item{service}{the EMODnet OGC WFS service name. Defaults to the EMODnet Seabed
Habitats WFS `"seabed_habitats_individual_habitat_map_and_model_datasets"`. For
available services, see `emodnet_wfs`.}

\item{service_version}{the WFS service version. Defaults to "2.0.0".}

\item{attribute}{character string}
}
\value{
detailed summary of individual attributes (variable). Particularly useful for inspecting
factor or character variable levels or unique values.
}
\description{
Inspect layer attributes
}
\examples{
wfs <- EMODnetWFS::emodnet_init_wfs_client(service = "human_activities")
layer_attributes_get_name(wfs)
layer_attribute_inspect(wfs, layer = "maritimebnds", attribute = "sitename")
}
