% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_attributes.R
\name{layer_attributes_get_names}
\alias{layer_attributes_get_names}
\title{Get names of layer attributes}
\usage{
layer_attributes_get_names(
  wfs = NULL,
  service = NULL,
  service_version = "2.0.0",
  layer
)
}
\arguments{
\item{wfs}{A \code{WFSClient} R6 object with methods for interfacing an OGC Web Feature Service.}

\item{service}{the EMODnet OGC WFS service name. For
available services, see \code{emodnet_wfs}.}

\item{service_version}{the WFS service version. Defaults to "2.0.0".}

\item{layer}{character sting of layer name. To get info on layers, including
\code{layer_name} use \code{\link[=emodnet_get_wfs_info]{emodnet_get_wfs_info()}}.}
}
\value{
character vector of layer attribute (variable) names.
}
\description{
Get names of layer attributes
}
\examples{
layer_attributes_get_names(service = "human_activities", layer = "maritimebnds")
}
