% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_attributes.R
\name{layer_attributes_summarise}
\alias{layer_attributes_summarise}
\title{Get summaries of layer attributes (variables)}
\usage{
layer_attributes_summarise(
  layer,
  wfs = NULL,
  service,
  service_version = "2.0.0"
)
}
\arguments{
\item{layer}{character sting of layer name. To get info on layers, including
\code{layer_name} use \code{\link[=emodnet_get_wfs_info]{emodnet_get_wfs_info()}}.}

\item{wfs}{A \code{WFSClient} R6 object with methods for interfacing an OGC Web Feature Service.}

\item{service}{the EMODnet OGC WFS service name. Defaults to the EMODnet Seabed
Habitats WFS \code{"seabed_habitats_individual_habitat_map_and_model_datasets"}. For
available services, see \code{emodnet_wfs}.}

\item{service_version}{the WFS service version. Defaults to "2.0.0".}
}
\value{
output of \code{summary()} on the attributes (variables) in a given layer for a given service.
}
\description{
Get summaries of layer attributes (variables)
}
\examples{
layer_attributes_summarise(service = "human_activities", layer = "maritimebnds")
}
