% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-roxygen2.R
\name{should_run_example}
\alias{should_run_example}
\title{Should we run this emodnet.wfs example?}
\usage{
should_run_example()
}
\value{
A Boolean indicating whether to run the example.
}
\description{
Determines whether we are in pkgdown or in an interactive session.
Do not use it in your own package!
}
\examples{
should_run_example()
}
