% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{emodnet_get_layers}
\alias{emodnet_get_layers}
\title{Get EMODnet WFS layers}
\usage{
emodnet_get_layers(
  wfs = NULL,
  service = "seabed_habitats_individual_habitat_map_and_model_datasets",
  service_version = "2.0.0",
  layers,
  crs = 4326,
  reduce_layers = FALSE,
  suppress_warnings = FALSE
)
}
\arguments{
\item{wfs}{A `WFSClient` R6 object with methods for interfacing an OGC Web Feature Service.}

\item{service}{the EMODnet OGC WFS service name. Defaults to the EMODnet Seabed
Habitats WFS `"seabed_habitats_individual_habitat_map_and_model_datasets"`. For
available services, see `emodnet_wfs`.}

\item{service_version}{the WFS service version. Defaults to "2.0.0".}

\item{layers}{a character vector of layer names. To get info on layers, including
`layer_name` use `emodnet_get_wfs_info()`}

\item{crs}{integer. An EPSG code for the outpur crs. Defaults to 4326, corresponsing to `"+proj=longlat +datum=WGS84 +no_defs"`}

\item{reduce_layers}{whether to reduce output layers to a single `sf` object.}

\item{suppress_warnings}{logical. Whether to suppress messages of layer
download failures.}
}
\value{
If `reduce_layers = FALSE` (default), a list of `sf`
objects, one element for each layer. Any layers for which download was
unsuccessful will be NULL. If `reduce_layers = TRUE`, all layers are
reduced (if possible) to a single `sf` containing data for all layers.
`NULL` layers are ignored.
}
\description{
Get EMODnet WFS layers
}
\examples{
emodnet_get_layers(layers = c("dk003069", "dk003070"))
emodnet_get_layers(layers = c("dk003069", "dk003070"), reduce_layers = TRUE)
}
