% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{emodnet_get_layers}
\alias{emodnet_get_layers}
\title{Get EMODnet WFS layers}
\usage{
emodnet_get_layers(
  wfs = NULL,
  service = NULL,
  service_version = NULL,
  layers,
  crs = NULL,
  cql_filter = NULL,
  reduce_layers = FALSE,
  ...
)
}
\arguments{
\item{wfs}{A \code{WFSClient} R6 object with methods for interfacing an
OGC Web Feature Service.}

\item{service}{the EMODnet OGC WFS service name.
For available services, see \code{\link[=emodnet_wfs]{emodnet_wfs()}}.}

\item{service_version}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
the WFS service version. Now always "2.0.0".}

\item{layers}{a character vector of layer names. To get info on layers,
including
\code{layer_name} use \code{\link[=emodnet_get_wfs_info]{emodnet_get_wfs_info()}}.}

\item{crs}{integer. EPSG code for the output crs. If \code{NULL} (default), layers
are returned with original crs.}

\item{cql_filter}{character. Features returned can be filtered using valid
Extended Common Query Language (ECQL) filtering statements
(\url{https://docs.geoserver.org/stable/en/user/filter/ecql_reference.html}).
Should be one of:
\itemize{
\item{character string or character vector of length 1.
Filter will be recycled across all layers requested}
\item{character vector of length equal to the length of layers.
Filter will be matched to layers sequentially.
Elements containing \code{NA} are ignored}
\item{named character vector. Each filter will be applied to the layer
corresponding to the filter name.
Filters with names that do not correspond to any layers are ignored.
Layers without corresponding filters are returned whole }
}}

\item{reduce_layers}{whether to reduce output layers to a single \code{sf} object.}

\item{...}{additional vendor parameter arguments passed to
\href{https://docs.geoserver.org/stable/en/user/services/wfs/reference.html#getfeature}{\code{ows4R::GetFeature()}}.# nolint
For example, including \code{count=1} returns the first available feature.}
}
\value{
If \code{reduce_layers = FALSE} (default), a list of \code{sf}
objects, one element for each layer. Any layers for which download was
unsuccessful will be NULL. If \code{reduce_layers = TRUE}, all layers are
reduced (if possible) to a single \code{sf} containing data for all layers.
\code{NULL} layers are ignored. \code{reduce_layers = TRUE} can also be used to return
an \code{sf} out of a single layer request instead of a list of length 1.
}
\description{
Performs an WFS getFeature request for layers from a \code{wfs} object or
specified EMODnet Service. Filtering
of layer features can also be handled via ECQL language filters.
}
\examples{
\dontrun{
# Layers as character vector
emodnet_get_layers(
  service = "biology",
  layers = c("mediseh_zostera_m_pnt", "mediseh_posidonia_nodata")
)


# Usage of cql_filter
emodnet_get_layers(
  service = "biology",
  layers = "mediseh_zostera_m_pnt",
  cql_filter = "country = 'Francia'"
)
# Usage of vendor parameter
emodnet_get_layers(
  service = "biology",
  layers = "mediseh_zostera_m_pnt",
  count = 1
)
}
}
