% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{emodnet_get_layers}
\alias{emodnet_get_layers}
\title{Get EMODnet WFS layers}
\usage{
emodnet_get_layers(
  wfs = NULL,
  service = "https://ows.emodnet-seabedhabitats.eu/emodnet_open_maplibrary/wfs",
  service_version = "2.0.0",
  layers,
  reduce_layers = FALSE
)
}
\arguments{
\item{wfs}{A `WFSClient` R6 object with methods for interfacing an OGC Web Feature Service.}

\item{service}{the EMODnet OGC WFS service URL. Defaults to the EMODnet Seabed
Habitats WFS.}

\item{service_version}{the WFS service version. Defaults to "2.0.0".}

\item{layers}{a character vector of layer names. To get info on layers, including
`layer_name` use `emodnet_get_wfs_info()`}

\item{reduce_layers}{whether to reduce output layers to a single `sf` object.}
}
\value{
either a single `sf` containing all layers (default) or a list of `sf`
objects, one for each layer (if `reduce_layers = TRUE`)
}
\description{
Get EMODnet WFS layers
}
\examples{
emodnet_get_layers(layers = c("dk003069", "dk003070"))
emodnet_get_layers(layers = c("dk003069", "dk003070"), reduce_layers = TRUE)
}
