% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{emodnet_get_wfs_info}
\alias{emodnet_get_wfs_info}
\title{Get WFS available layer information}
\usage{
emodnet_get_wfs_info(
  wfs = NULL,
  service = "https://ows.emodnet-seabedhabitats.eu/emodnet_open_maplibrary/wfs",
  service_version = "2.0.0"
)
}
\arguments{
\item{wfs}{A `WFSClient` R6 object with methods for interfacing an OGC Web Feature Service.}

\item{service}{the EMODnet OGC WFS service URL. Defaults to the EMODnet Seabed
Habitats WFS.}

\item{service_version}{the WFS service version. Defaults to "2.0.0".}
}
\value{
a tibble summarising information on each layer available from the service
}
\description{
Get WFS available layer information
}
\examples{
emodnet_get_wfs_info()

}
