% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{emodnet_get_wfs_info}
\alias{emodnet_get_wfs_info}
\title{Get WFS available layer information}
\usage{
emodnet_get_wfs_info(
  wfs = NULL,
  service = "seabed_habitats_individual_habitat_map_and_model_datasets",
  service_version = "2.0.0"
)
}
\arguments{
\item{wfs}{A `WFSClient` R6 object with methods for interfacing an OGC Web Feature Service.}

\item{service}{the EMODnet OGC WFS service name. Defaults to the EMODnet Seabed
Habitats WFS `"seabed_habitats_individual_habitat_map_and_model_datasets"`. For
available services, see `emodnet_wfs`.}

\item{service_version}{the WFS service version. Defaults to "2.0.0".}
}
\value{
a tibble summarising information on each layer available from the service
}
\description{
Get WFS available layer information
}
\examples{
emodnet_get_wfs_info()

}
