% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emodnet_wfs.R
\name{emodnet_wfs}
\alias{emodnet_wfs}
\title{Available EMODnet Web Feature Services}
\usage{
emodnet_wfs()
}
\value{
Tibble of available EMODnet Web Feature Services
}
\description{
Available EMODnet Web Feature Services
}
