% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emodnet_wfs.R
\name{emodnet_wfs}
\alias{emodnet_wfs}
\title{Which data sources (services) are available?}
\format{
\subsection{\code{emodnet_wfs}}{

\describe{
\item{emodnet_thematic_lot}{EMODnet disciplinary themes - bathymetry,
biology, chemistry, geology, human activities, physics and seabed habitats}
\item{service_name}{Name of the specific service.
Use in \link{emodnet_init_wfs_client}.}
\item{service_url}{
\href{https://www.ogc.org/publications/standard/wfs/}{Web Feature Service (WFS)}
URL endpoint for accessing the service.}
}
}
}
\usage{
emodnet_wfs()
}
\value{
Tibble of available EMODnet Web Feature Services
}
\description{
Available EMODnet Web Feature Services
}
\examples{
\dontshow{if (identical(Sys.getenv("IN_PKGDOWN"), "true") || interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
emodnet_wfs()
\dontshow{\}) # examplesIf}
}
