% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emodnet_wfs.R
\name{emodnet_wfs}
\alias{emodnet_wfs}
\title{Which data sources (services) are available?}
\usage{
emodnet_wfs()
}
\value{
Tibble of available EMODnet Web Feature Services
}
\description{
Available EMODnet Web Feature Services
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
emodnet_wfs()
\dontshow{\}) # examplesIf}
}
