% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_attributes.R
\name{layer_attribute_inspect}
\alias{layer_attribute_inspect}
\title{Inspect layer attribute}
\usage{
layer_attribute_inspect(
  wfs = NULL,
  service = NULL,
  service_version = "2.0.0",
  layer,
  attribute
)
}
\arguments{
\item{wfs}{A \code{WFSClient} R6 object with methods for interfacing an OGC Web Feature Service.}

\item{service}{the EMODnet OGC WFS service name.
For available services, see \code{\link[=emodnet_wfs]{emodnet_wfs()}}.}

\item{service_version}{the WFS service version. Defaults to "2.0.0".}

\item{layer}{character sting of layer name. To get info on layers, including
\code{layer_name} use \code{\link[=emodnet_get_wfs_info]{emodnet_get_wfs_info()}}.}

\item{attribute}{character string, name of layer attribute (variable). Use
\code{\link[=layer_attributes_get_names]{layer_attributes_get_names()}} to get layer attribute names.}
}
\value{
Detailed summary of individual attribute (variable). Particularly useful for inspecting
factor or character variable levels or unique values.
}
\description{
Inspect layer attribute
}
\examples{
\dontrun{
wfs <- emodnet_init_wfs_client(service = "human_activities")
layer_attributes_get_names(wfs, layer = "maritimebnds")
layer_attribute_inspect(wfs, layer = "maritimebnds", attribute = "sitename")
}
}
