% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_attributes.R
\name{layer_attributes_summarise}
\alias{layer_attributes_summarise}
\title{Get summaries of layer attributes (variables)}
\usage{
layer_attributes_summarise(
  wfs = NULL,
  service = NULL,
  service_version = NULL,
  layer
)
}
\arguments{
\item{wfs}{A \code{WFSClient} R6 object with methods for interfacing an
OGC Web Feature Service. From \code{\link[=emodnet_init_wfs_client]{emodnet_init_wfs_client()}}.}

\item{service}{the EMODnet OGC WFS service name.
For available services, see \code{\link[=emodnet_wfs]{emodnet_wfs()}}.}

\item{service_version}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
the WFS service version. Now always "2.0.0".}

\item{layer}{character sting of layer name. To get info on layers, including
\code{layer_name} use \code{\link[=emodnet_get_wfs_info]{emodnet_get_wfs_info()}}.}
}
\value{
output of \code{summary()} on the attributes (variables) in a given layer
for a given service.
}
\description{
Get summaries of layer attributes (variables)
}
\examples{
\dontrun{
layer_attributes_summarise(
  service = "biology",
  layer = "mediseh_zostera_m_pnt"
)
}
}
